+++ src/slack_mcp/tools/core.py
@@ -135,231 +135,6 @@ def register_tools(mcp):
         except Exception as e:
             return f"❌ Unexpected Error: {str(e)}"
 
-    @mcp.tool("get_file_content")
-    async def get_file_content(
-        file_id: str,
-        max_size_mb: float = 10
-    ) -> str:
-        """
-        Get the content of a text file from Slack (download functionality).
-        
-        Args:
-            file_id: The ID of the file to download
-            max_size_mb: Maximum file size to download in MB (default: 10MB)
-        
-        Returns:
-            The file content or error message
-        """
-        try:
-            validate_slack_token()
-            
-            # Initialize async client
-            async_client = await init_async_client()
-            if not async_client:
-                return "❌ Failed to initialize async Slack client"
-            
-            # Get file info first
-            info_response = await make_slack_request(
-                async_client.files_info,
-                file=file_id
-            )
-            
-            if not info_response:
-                return "❌ File not found or inaccessible"
-            
-            file_info = info_response.get('file', {})
-            file_name = file_info.get('name', 'Unknown')
-            file_size = file_info.get('size', 0)
-            mimetype = file_info.get('mimetype', '')
-            
-            # Check file size
-            max_size_bytes = max_size_mb * 1024 * 1024
-            if file_size > max_size_bytes:
-                return f"❌ File too large: {file_size / (1024 * 1024):.1f}MB exceeds {max_size_mb}MB limit"
-            
-            # Check if it's a text-based file
-            text_mimetypes = ['text/', 'application/json', 'application/xml', 'application/javascript', 
-                             'application/x-yaml', 'application/x-python', 'application/x-sh']
-            
-            is_text_file = any(mimetype.startswith(mt) for mt in text_mimetypes)
-            
-            if not is_text_file and not file_info.get('filetype') in ['text', 'json', 'xml', 'yaml', 'python', 'javascript', 'sh', 'bash']:
-                return f"❌ File appears to be binary (MIME: {mimetype}). This tool only supports text files."
-            
-            # Get download URL
-            download_url = file_info.get('url_private_download')
-            if not download_url:
-                return "❌ No download URL available for this file"
-            
-            # Download the file content
-            import httpx
-            headers = {
-                'Authorization': f'Bearer {async_client.token}',
-                'User-Agent': MCP_USER_AGENT
-            }
-            
-            async with httpx.AsyncClient() as client:
-                response = await client.get(download_url, headers=headers)
-                
-                if response.status_code == 200:
-                    content = response.text
-                    
-                    result = f"📄 **File Content Retrieved**\n\n"
-                    result += f"**File:** {file_name} (ID: {file_id})\n"
-                    result += f"**Size:** {len(content)} characters\n"
-                    result += f"**Type:** {mimetype or file_info.get('filetype', 'unknown')}\n\n"
-                    result += "**Content:**\n```\n"
-                    
-                    # Limit displayed content if too long
-                    if len(content) > 50000:
-                        result += content[:50000]
-                        result += "\n... (content truncated at 50,000 characters)\n"
-                    else:
-                        result += content
-                    
-                    result += "\n```"
-                    return result
-                else:
-                    return f"❌ Failed to download file: HTTP {response.status_code}"
-                    
-        except ValueError as e:
-            return f"❌ Configuration Error: {str(e)}"
-        except SlackApiError as e:
-            return f"❌ Slack API Error: {e.response['error']}"
-        except Exception as e:
-            return f"❌ Unexpected Error: {str(e)}"
-
-    @mcp.tool("list_files")
-    async def list_files(
-        channel: Optional[str] = None,
-        user: Optional[str] = None,
-        types: Optional[str] = None,
-        count: int = 20,
-        page: int = 1
-    ) -> str:
-        """
-        List files in the Slack workspace with optional filtering.
-        
-        Args:
-            channel: Channel ID to filter files from
-            user: User ID to filter files uploaded by
-            types: Comma-separated list of file types (e.g., 'images,pdfs,zips')
-            count: Number of files to return per page (default: 20, max: 100)
-            page: Page number of results to return (default: 1)
-        
-        Returns:
-            Formatted list of files with details
-        """
-        try:
-            validate_slack_token()
-            
-            # Initialize async client
-            async_client = await init_async_client()
-            if not async_client:
-                return "❌ Failed to initialize async Slack client"
-            
-            # Build request parameters
-            params = {
-                "count": min(count, 100),  # Cap at 100
-                "page": page
-            }
-            
-            if channel:
-                params["channel"] = channel
-            if user:
-                params["user"] = user
-            if types:
-                params["types"] = types
-            
-            # List files with rate limiting
-            response = await make_slack_request(
-                async_client.files_list,
-                **params
-            )
-            
-            if response:
-                files = response.get('files', [])
-                total = response.get('paging', {}).get('total', 0)
-                pages = response.get('paging', {}).get('pages', 1)
-                current_page = response.get('paging', {}).get('page', 1)
-                
-                result = f"📁 **Files in Workspace** (Page {current_page}/{pages}, Total: {total})\n\n"
-                
-                if not files:
-                    result += "No files found matching the criteria.\n"
-                    return result
-                
-                # Group files by type
-                images = []
-                documents = []
-                videos = []
-                other = []
-                
-                for file in files:
-                    if file.get('mimetype', '').startswith('image/'):
-                        images.append(file)
-                    elif file.get('mimetype', '').startswith(('text/', 'application/pdf', 'application/msword', 'application/vnd.')):
-                        documents.append(file)
-                    elif file.get('mimetype', '').startswith('video/'):
-                        videos.append(file)
-                    else:
-                        other.append(file)
-                
-                # Import formatting helper
-                from ..utils.formatting import format_file_info
-                
-                # Format file listings by type
-                if images:
-                    result += f"**🖼️ Images ({len(images)}):**\n"
-                    for i, file in enumerate(images[:5], 1):
-                        result += format_file_info(file, i)
-                    if len(images) > 5:
-                        result += f"... and {len(images) - 5} more images\n"
-                    result += "\n"
-                
-                if documents:
-                    result += f"**📄 Documents ({len(documents)}):**\n"
-                    for i, file in enumerate(documents[:5], 1):
-                        result += format_file_info(file, i)
-                    if len(documents) > 5:
-                        result += f"... and {len(documents) - 5} more documents\n"
-                    result += "\n"
-                
-                if videos:
-                    result += f"**🎥 Videos ({len(videos)}):**\n"
-                    for i, file in enumerate(videos[:3], 1):
-                        result += format_file_info(file, i)
-                    if len(videos) > 3:
-                        result += f"... and {len(videos) - 3} more videos\n"
-                    result += "\n"
-                
-                if other:
-                    result += f"**📎 Other Files ({len(other)}):**\n"
-                    for i, file in enumerate(other[:5], 1):
-                        result += format_file_info(file, i)
-                    if len(other) > 5:
-                        result += f"... and {len(other) - 5} more files\n"
-                    result += "\n"
-                
-                # Add pagination info
-                if pages > 1:
-                    result += f"\n**Navigation:** Page {current_page} of {pages}\n"
-                    if current_page > 1:
-                        result += f"• Previous page: use page={current_page - 1}\n"
-                    if current_page < pages:
-                        result += f"• Next page: use page={current_page + 1}\n"
-                
-                return result
-            else:
-                return "❌ Failed to list files: API request failed"
-                
-        except ValueError as e:
-            return f"❌ Configuration Error: {str(e)}"
-        except SlackApiError as e:
-            return f"❌ Slack API Error: {e.response['error']}"
-        except Exception as e:
-            return f"❌ Unexpected Error: {str(e)}"
-
     @mcp.tool("send_message")
     async def send_message(
         channel: str,
@@ -803,101 +578,3 @@ def register_tools(mcp):
         except Exception as e:
             return f"❌ Unexpected Error: {str(e)}"
 
-    @mcp.tool("upload_file")
-    async def upload_file(
-        channels: str,
-        content: Optional[str] = None,
-        filename: Optional[str] = None,
-        filetype: Optional[str] = None,
-        title: Optional[str] = None,
-        initial_comment: Optional[str] = None
-    ) -> str:
-        """
-        Upload a file to one or more Slack channels or users.
-        
-        Args:
-            channels: Comma-separated list of channel IDs, channel names, or user IDs
-            content: File content as text (for text files)
-            filename: Name of the file
-            filetype: File type (e.g., 'text', 'json', 'csv', 'python', 'javascript')
-            title: Title of the file
-            initial_comment: Initial comment for the file
-        
-        Returns:
-            File upload result with file ID and sharing info
-        """
-        try:
-            validate_slack_token()
-            
-            if not content:
-                return "❌ File content is required"
-            
-            # Initialize async client
-            async_client = await init_async_client()
-            if not async_client:
-                return "❌ Failed to initialize async Slack client"
-            
-            # Use files_upload_v2 with rate limiting
-            response = await make_slack_request(
-                async_client.files_upload_v2,
-                content=content,
-                filename=filename or "file.txt",
-                title=title,
-                initial_comment=initial_comment,
-                channel=channels
-            )
-            
-            if response:
-                file_info = response["file"]
-                
-                result = "✅ **File Uploaded Successfully**\n\n"
-                result += f"• **File ID:** {file_info.get('id', 'N/A')}\n"
-                result += f"• **Name:** {file_info.get('name', 'N/A')}\n"
-                
-                if file_info.get('title'):
-                    result += f"• **Title:** {file_info['title']}\n"
-                
-                # File size formatting
-                if file_info.get('size'):
-                    size = file_info['size']
-                    if size < 1024:
-                        size_str = f"{size} bytes"
-                    elif size < 1024 * 1024:
-                        size_str = f"{size / 1024:.1f} KB"
-                    else:
-                        size_str = f"{size / (1024 * 1024):.1f} MB"
-                    result += f"• **Size:** {size_str}\n"
-                
-                if file_info.get('mimetype'):
-                    result += f"• **Type:** {file_info['mimetype']}\n"
-                elif filetype:
-                    result += f"• **Type:** {filetype}\n"
-                
-                # Channels the file was shared to
-                result += f"\n• **Shared to:** {channels}\n"
-                
-                if initial_comment:
-                    result += f"• **Comment:** {initial_comment}\n"
-                
-                # File URLs
-                if file_info.get('url_private'):
-                    result += f"\n• **URL:** {file_info['url_private']}\n"
-                
-                if file_info.get('permalink'):
-                    result += f"• **Permalink:** {file_info['permalink']}\n"
-                
-                # Upload timestamp
-                if file_info.get('created'):
-                    upload_time = datetime.fromtimestamp(file_info['created']).strftime('%Y-%m-%d %H:%M:%S')
-                    result += f"\n• **Uploaded at:** {upload_time}\n"
-                
-                return result
-            else:
-                return "❌ Failed to upload file: API request failed"
-                
-        except ValueError as e:
-            return f"❌ Configuration Error: {str(e)}"
-        except SlackApiError as e:
-            return f"❌ Slack API Error: {e.response['error']}"
-        except Exception as e:
-            return f"❌ Unexpected Error: {str(e)}"
